% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{spe_vignette}
\alias{spe_vignette}
\title{Mock SpatialExperiment for Vignettes}
\format{
A SpatialExperiment object
}
\source{
Internal simulation
}
\description{
A minimal SpatialExperiment object used for demonstrating
the spatial artifact detection workflow.
}
