% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{spe_vignette}
\alias{spe_vignette}
\title{Mock SpatialExperiment for Vignettes}
\format{
A \code{\link[SpatialExperiment]{SpatialExperiment}} object.
}
\source{
Internal simulation
}
\value{
A \code{\link[SpatialExperiment]{SpatialExperiment}} object.
}
\description{
A minimal \code{\link[SpatialExperiment]{SpatialExperiment}} object used for
demonstrating the spatial artifact detection workflow.
}
\examples{
data(spe_vignette)
spe_vignette
}
